requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
backports.zoneinfo==0.2.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-webpush==0.3.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
http_ece==1.2.0
idna==3.7
multidict==6.0.5
pillow==10.4.0
pip==24.0
py-vapid==1.9.1
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.1
tzdata==2024.1
typing_extensions==4.7.1
pytz==2024.1
setuptools==47.1.0
pywebpush==2.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
